Patch 0362
・メモリマネージャにおいて、ユニットが全て埋まって空きが無いブロックに対してメモリリーク検出が
　行われない潜在的なバグを修正（malloc.c）
	unfill_prevがNULLのときは、「ブロックが空っぽ」「ブロックが満タン」の2つの状態の
	いずれかを意味します。
	したがってunfill_prevではブロックにまだユニットが残っているかどうかを判定すること
	はできません。
	代わりにunit_usedを見るように修正しています。
・timer.cの細かい整形（timer.c）
・一部の構造体のメンバを見直し（timer.h, battle.c, map.h, skill.*, status.*）

+++ src/map/battle.c
@@ -35,8 +35,8 @@ struct battle_delay_damage_ {
 	struct block_list *src;
 	int target;
 	int damage;
-	int skillid;
-	int skilllv;
+	short skillid;
+	short skilllv;
 	int flag;
 };
